ocr/detection/east/requirements-preprocess.txt
shapely
Pillow
numpy
h5py
opencv-python